Skip pg_semver for PGv16

+++ docker-compose/ext-src/pg_semver-src/regular-test.sh
@@ -1,5 +1,10 @@
-#!/bin/sh
+#!/bin/bash
 set -ex
+# For v16 it's required to create a type which is impossible without superuser access
+# do not run this test so far
+if [[ "${PG_VERSION}" = v16 ]]; then
+  exit 0
+fi
 cd "$(dirname ${0})"
 dropdb --if-exist contrib_regression
 createdb contrib_regression
